Switch mlir_canonicalize to generic pass runner, reorganize driver files
