+++ mlir/include/Quantum-c/Dialects.h
@@ -30,9 +30,12 @@ enum CatalystCReturnCode {
     ReturnTranslationFailed
 };
 
-/// Canonicalize an MLIR module, given in textual form. This performs a round-trip, allocating a
-/// char buffer in "dest" that the caller must take ownership of freeing.
-CatalystCReturnCode Canonicalize(const char *source, char **dest);
+/// Run a given set of passes on an MLIR module.
+///
+/// The IR is supplied in textual form while the passes are expected in MLIR's command line
+/// interface form. This allocates a buffer for the resulting textual IR that the caller must
+/// take ownership of freeing.
+CatalystCReturnCode RunPassPipeline(const char *source, const char *passes, char **dest);
 
 /// Entry point to the MLIR portion of the compiler.
 CatalystCReturnCode QuantumDriverMain(const char *source, bool keepIntermediate = false);
