+++ mlir/lib/CAPI/CMakeLists.txt
@@ -3,6 +3,7 @@ add_mlir_public_c_api_library(QuantumCAPI
     Dialects.cpp
 
     LINK_LIBS PRIVATE
+    CatalystCompilerDriver
     MLIRCatalyst
     MLIRCatalystTransforms
     MLIRQuantum
