+++ mlir/lib/Catalyst/CMakeLists.txt
@@ -1,2 +1,3 @@
+add_subdirectory(Driver)
 add_subdirectory(IR)
 add_subdirectory(Transforms)
