+++ mlir/lib/Catalyst/Driver/CMakeLists.txt
@@ -1,0 +1,6 @@
+add_mlir_library(CatalystCompilerDriver
+    LLVMTarget.cpp
+    Pipelines.cpp
+
+    LINK_LIBS PRIVATE
+)
